Add missing bindings

+++ jolt/core/src/main/java/jolt/BodyActivationListenerEm.java
@@ -42,20 +42,20 @@ public class BodyActivationListenerEm extends BodyActivationListener {
     private void setupCallback() {
     }
 
-    protected void OnBodyActivated(BodyID inBodyID, int inBodyUserData) {
+    protected void OnBodyActivated(BodyID inBodyID, long inBodyUserData) {
     }
 
-    private void internal_OnBodyActivated(long inBodyID_addr, int inBodyUserData) {
+    private void internal_OnBodyActivated(long inBodyID_addr, long inBodyUserData) {
         if (BodyID_TEMP_STATIC_GEN_0 == null)
             BodyID_TEMP_STATIC_GEN_0 = BodyID.native_new();
         BodyID_TEMP_STATIC_GEN_0.internal_reset(inBodyID_addr, false);
         OnBodyActivated(BodyID_TEMP_STATIC_GEN_0, inBodyUserData);
     }
 
-    protected void OnBodyDeactivated(BodyID inBodyID, int inBodyUserData) {
+    protected void OnBodyDeactivated(BodyID inBodyID, long inBodyUserData) {
     }
 
-    private void internal_OnBodyDeactivated(long inBodyID_addr, int inBodyUserData) {
+    private void internal_OnBodyDeactivated(long inBodyID_addr, long inBodyUserData) {
         if (BodyID_TEMP_STATIC_GEN_1 == null)
             BodyID_TEMP_STATIC_GEN_1 = BodyID.native_new();
         BodyID_TEMP_STATIC_GEN_1.internal_reset(inBodyID_addr, false);
