+++ jolt/core/src/main/java/jolt/physics/constraints/Constraint.java
@@ -79,11 +79,11 @@ public class Constraint extends NativeObject {
         return false;
     }
 
-    public int GetUserData() {
+    public long GetUserData() {
         return 0;
     }
 
-    public void SetUserData(int inUserData) {
+    public void SetUserData(long inUserData) {
     }
 
     public void ResetWarmStart() {
