+++ jolt/shared/jni/src/main/java/jolt/physics/constraints/natives/JNI_Constraint.java
@@ -38,9 +38,9 @@ final public class JNI_Constraint {
 
     public static native boolean internal_native_IsActive(long this_addr);
 
-    public static native int internal_native_GetUserData(long this_addr);
+    public static native long internal_native_GetUserData(long this_addr);
 
-    public static native void internal_native_SetUserData(long this_addr, int inUserData);
+    public static native void internal_native_SetUserData(long this_addr, long inUserData);
 
     public static native void internal_native_ResetWarmStart(long this_addr);
