+++ jolt/shared/jni/src/main/java/jolt/physics/softbody/natives/JNI_SoftBodyCreationSettings.java
@@ -22,9 +22,9 @@ final public class JNI_SoftBodyCreationSettings {
 
     public static native void internal_native_set_mRotation(long this_addr, long mRotation_addr);
 
-    public static native int internal_native_get_mUserData(long this_addr);
+    public static native long internal_native_get_mUserData(long this_addr);
 
-    public static native void internal_native_set_mUserData(long this_addr, int mUserData);
+    public static native void internal_native_set_mUserData(long this_addr, long mUserData);
 
     public static native int internal_native_get_mObjectLayer(long this_addr);
