+++ samples/shared/src/main/java/jolt/example/samples/app/tests/raycast/RayCastData.java
@@ -3,6 +3,6 @@ package jolt.example.samples.app.tests.raycast;
 import com.badlogic.gdx.math.Vector3;
 
 public class RayCastData {
-    public int userData;
+    public long userData;
     public final Vector3 hitPosition = new Vector3();
 }
